+++ R/meta-internal.R
@@ -126,6 +126,13 @@ augment <- function(x, len, fun) {
 }
 
 
+stoponly <- function(arg, val, func)
+  stop("Argument ", arg, " =\"", val, "\"",
+               " only defined for meta-analysis conducted with ",
+               func, ".",
+               call. = FALSE)
+
+
 .settings <- list()
 ##
 ## Set defaults (for internal options)
@@ -150,8 +157,10 @@ setOption("meth4tau", c("DL", "PM", "REML", "ML", "HS", "SJ", "HE", "EB"))
 setOption("meth4tau.ci", c("QP", "BJ", "J", "PL", ""))
 setOption("adhoc4hakn", c("", "se", "ci", "iqwig6"))
 ##
-setOption("meth4bias", c("rank", "linreg", "mm", "count", "score",
-                         "peters", "deeks"))
+setOption("meth4bias.old", c("rank", "linreg", "mm", "count", "score"))
+setOption("meth4bias", c("Begg", "Egger", "Thompson", "Schwarzer",
+                         "Harbord", "Peters", "Deeks",
+                         "Pustejovsky", "Macaskill"))
 ##
 ## List of arguments that can be changed by user
 ##
@@ -197,7 +206,7 @@ setOption("method.tau.ci", NULL)
 setOption("tau.common", FALSE)
 setOption("prediction", FALSE)
 setOption("level.predict", 0.95)
-setOption("method.bias", "linreg")
+setOption("method.bias", "Egger")
 setOption("text.fixed", "Fixed effect model")
 setOption("text.random", "Random effects model")
 setOption("text.predict", "Prediction interval")